HumanDetectionクラスとHumanFollowクラスに分割後、リファクタリング
