+++ tang_detection/launch/tang_detection.launch
@@ -6,7 +6,7 @@
     <param name="port" value="$(arg port)" />
   </node>
 
-  <node name="tang_detection" pkg="tang_detection" type="human_detection.py" output="screen">
+  <node name="tang_detection" pkg="tang_detection" type="human_follow.py" output="screen">
     <param name = "threshold" type = "double" value = "6.0"/>  <!-- これより遠い距離の画素を無視する -->
     <param name = "debug" type="bool" value="1"/>
   </node>
