+++ tang_detection/scripts/config.py
@@ -28,3 +28,7 @@ class GearConfig:
     Pitch = 6.1850105367549055/2/1000
     right_gear_pin = 24
     left_gear_pin = 22
+
+class HumanDetectionConfig:
+    # ssd-mobilenet-v2, peoplenet
+    model = "ssd-mobilenet-v2"
